feat: reduce deps and collect in sel4-runtime crate
