+++ services/fat-thread/Cargo.toml
@@ -5,7 +5,6 @@ edition = "2024"
 
 [dependencies]
 sel4 = { workspace = true }
-sel4-logging = { workspace = true }
 sel4-runtime = { workspace = true }
 spin = { workspace = true }
 log = "0.4"
