+++ services/uart-thread/src/main.rs
@@ -20,56 +20,43 @@ sel4_runtime::entry_point!(main);
 // 4. 利用 rust 的 trait object 来实现 IPC 的分发
 
 fn main() -> ! {
-    common::init_log!(log::LevelFilter::Debug);
-    common::init_recv_slot();
-
     log::info!("Booting...");
     // let mut pl011 = Pl011UartIfaceImpl::new(VIRTIO_MMIO_VIRT_ADDR);
     let mut pl011 = PL011DRV.lock();
 
-    loop {
-        let (msg, badge) = DEFAULT_SERVE_EP.recv(());
-        log::warn!("recv msg: {:?}", msg);
-        match badge {
-            // u64::MAX => irq_callback(),
-            _ => {
-                let msg_label = match UartIfaceEvent::try_from(msg.label()) {
-                    Ok(label) => label,
-                    Err(_) => continue,
-                };
-                match msg_label {
-                    UartIfaceEvent::init => {
-                        with_ipc_buffer_mut(|ib| {
+    with_ipc_buffer_mut(|ib| {
+        loop {
+            let (msg, badge) = DEFAULT_SERVE_EP.recv(());
+            log::warn!("recv msg: {:?}", msg);
+            match badge {
+                // u64::MAX => irq_callback(),
+                _ => {
+                    let msg_label = match UartIfaceEvent::try_from(msg.label()) {
+                        Ok(label) => label,
+                        Err(_) => continue,
+                    };
+                    match msg_label {
+                        UartIfaceEvent::init => {
                             sel4::reply(ib, MessageInfoBuilder::default().build());
-                        });
-                    }
-                    UartIfaceEvent::getchar => {
-                        let c = pl011.getchar();
-                        with_ipc_buffer_mut(|ib| {
+                        }
+                        UartIfaceEvent::getchar => {
+                            let c = pl011.getchar();
                             ib.msg_regs_mut()[0] = c as _;
                             sel4::reply(ib, MessageInfoBuilder::default().length(1).build());
-                        })
-                    }
-                    UartIfaceEvent::putchar => {
-                        with_ipc_buffer(|ib| {
+                        }
+                        UartIfaceEvent::putchar => {
                             pl011.putchar(ib.msg_bytes()[0]);
-                        });
-                        with_ipc_buffer_mut(|ib| {
                             sel4::reply(ib, MessageInfoBuilder::default().build());
-                        });
-                    }
-                    UartIfaceEvent::puts => {
-                        log::debug!("putstring");
-                        with_ipc_buffer(|ib| {
+                        }
+                        UartIfaceEvent::puts => {
+                            log::debug!("putstring");
                             let len = ib.msg_regs()[0] as usize;
                             pl011.puts(&ib.msg_bytes()[REG_LEN..len + REG_LEN]);
-                        });
-                        with_ipc_buffer_mut(|ib| {
                             sel4::reply(ib, MessageInfoBuilder::default().build());
-                        })
+                        }
                     }
                 }
             }
         }
-    }
+    })
 }
