+++ tasks/simple-cli/Cargo.toml
@@ -8,7 +8,6 @@ sel4 = { workspace = true }
 sel4-ctors-dtors = { workspace = true }
 sel4-panicking = { workspace = true, features = ["alloc"] }
 sel4-panicking-env = { workspace = true }
-sel4-logging = { workspace = true }
 sel4-dlmalloc = { workspace = true }
 sel4-sync = { workspace = true }
 sel4-runtime = { workspace = true }
