+++ tasks/test-demo/Cargo.toml
@@ -6,13 +6,12 @@ edition.workspace = true
 [features]
 uart-ipc = ["srv-iface/uart-ipc"]
 
-default = ["uart-ipc"]
+# default = ["uart-ipc"]
 
 [dependencies]
 log = "0.4"
 config = { workspace = true }
 sel4 = { workspace = true }
-sel4-logging = { workspace = true }
 common = { workspace = true }
 srv-iface = { workspace = true }
 sel4-runtime = { workspace = true }
